Moving test utilities for processing templates to test_utils plugin
