+++ projects/vdk-plugins/vdk-impala/tests/functional/jobs/load_dimension_scd1_template_job/02_run_load_dimension_scd1_template.py
@@ -1,6 +1,6 @@
 # Copyright 2021-2023 VMware, Inc.
 # SPDX-License-Identifier: Apache-2.0
-from tests.functional import test_utility
+from vdk.plugin.test_utils.processing_templates_utility import setup_testing_check
 from vdk.api.job_input import IJobInput
 
 __author__ = "VMware, Inc."
@@ -12,7 +12,7 @@ __copyright__ = (
 def run(job_input: IJobInput) -> None:
     template_args = job_input.get_arguments()
     check = template_args.get("check")
-    check = test_utility.setup_testing_check(check)
+    check = setup_testing_check(check)
 
     job_input.execute_template(
         template_name="load/dimension/scd1",
